# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  cluster.CreateClusterRequestBody:
    properties:
      cluster_name:
        type: string
      description:
        type: string
      project_name:
        type: string
    required:
    - cluster_name
    - project_name
    type: object
  cluster.UpdateClusterRequestBody:
    properties:
      cluster_name:
        type: string
      description:
        type: string
    type: object
...
  util.Res:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /cluster/create:
    post:
      consumes:
      - application/json
      description: Create Cluster
      operationId: CreateCluster
      parameters:
      - description: Create Cluster
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/cluster.CreateClusterRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: CreateCluster
      tags:
      - CreateCluster
  /cluster/delete:
    delete:
      consumes:
      - application/json
      description: Delete Cluster
      operationId: DeleteCluster
      parameters:
      - description: Project_Name
        in: query
        name: project_name
        required: true
        type: string
      - description: Cluster_Name
        in: query
        name: cluster_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: DeleteCluster
      tags:
      - DeleteCluster
  /cluster/get_id:
    get:
      consumes:
      - application/json
      description: Get ClusterID
      operationId: GetClusterId
      parameters:
      - description: Project Name
        in: query
        name: project_name
        required: true
        type: string
      - description: Cluster Name
        in: query
        name: cluster_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: GetClusterId
      tags:
      - GetClusterId
  /cluster/label:
    get:
      consumes:
      - application/json
      description: List Cluster for label
      operationId: ListGetLabel
      parameters:
      - description: Project Name
        in: query
        name: project_name
        type: string
      - description: Label Key
        in: query
        name: label_key
        required: true
        type: string
      - description: Label value
        in: query
        name: label_value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: ListGetLabel
      tags:
      - ListGetLabel
  /cluster/list:
    get:
      consumes:
      - application/json
      description: List  Cluster
      operationId: ListCluster
      parameters:
      - description: Project Name
        in: query
        name: project_name
        type: string
      - description: Cluster Name
        in: query
        name: cluster_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: ListCluster
      tags:
      - ListCluster
  /cluster/update:
    put:
      consumes:
      - application/json
      description: Update Cluster
      operationId: UpdateCluster
      parameters:
      - description: Project_Name
        in: query
        name: project_name
        required: true
        type: string
      - description: Cluster_ID
        in: query
        name: cluster_id
        required: true
        type: string
      - description: Update Cluster
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/cluster.UpdateClusterRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: UpdateCluster
      tags:
      - UpdateCluster
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"